Implemented Reference counter mechanism
vkw highly relies on usage of references to externally constructed
object. E.g. every device-based resource(like swapchain or semaphore)
need to store reference to device to be able to free itself in
destructor. For long time vkw expected applications to keep
device alive long enough enough for all such resources safely
to be freed, but that wasn't enforced. It may lead to device
be freed too early and all references to it dangle. Using
shared pointer here do not seem like good solution, because it
enforce that specific ownership strategy for all objects and
isn't intended. Better solution is actually to intrusively
track reference count of such object and make sure it is freed
or moved only if no references left. This avoids reference
dangling and do not enforce any ownership strategy to user.
However it adds additional overhead and is practically useless
if application is constructed error-free. That's why this solution
designed to be debug-only and can be easily switched off in
release configuration.